requirements.txt
Flask==2.2.3
Flask_Cors==3.0.10
nltk==3.8.1
Pillow==9.4.0
requests==2.28.1
torch==1.13.1
transformers==4.25.1
